// packages/mpl-core-kit-lib/src/generated/types/baseExtraAccount.ts
export type BaseExtraAccount =
  | {
      __kind: 'Address'
      address: Address
      isSigner: boolean
      isWritable: boolean
    }
  | {
      __kind: 'CustomPda'
      customProgramId: Option<Address>
      isSigner: boolean
      isWritable: boolean
      seeds: Array<BaseSeed>
    }
  | { __kind: 'PreconfiguredAsset'; isSigner: boolean; isWritable: boolean }
  | {
      __kind: 'PreconfiguredCollection'
      isSigner: boolean
      isWritable: boolean
    }
  | { __kind: 'PreconfiguredOwner'; isSigner: boolean; isWritable: boolean }
  | { __kind: 'PreconfiguredProgram'; isSigner: boolean; isWritable: boolean }
  | { __kind: 'PreconfiguredRecipient'; isSigner: boolean; isWritable: boolean }